diverse optimization steps

+++ bin/start-detector.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
-#touch juan-aaaaaaaaaaaaaaaaa.txt
-#result=`./objdetector -c=0.53 -d=$1 -t=$2`
 result=`./objectdetector $1 $2`
 
-# if [ "$result" == "1" ]; then
-# do something...
-# fi
+if [ "$result" == "detections=0" ]; then
 
+    file=$1
 
+    rm -f $file
+    
+    file=$2
+    
+    rm -f $file
+    
+    
+fi
 
-file=$1
-file=$file$"/image*.jpg"
-rm -f $file
 echo $result
